Serve images

+++ libs/inventory/src/write/infra/gateways/image-uploader/local-disk-storage-image-uploader.gateway.ts
@@ -41,6 +41,7 @@ export class LocalDiskStorageImageUploaderGateway
     const imagePath = fs.readdirSync(this.uploadFolderPath).find((file) => {
       return file.includes(imageId);
     });
+    if (!imagePath) return;
     fs.unlinkSync(path.join(this.uploadFolderPath, imagePath));
   }
 }
